Add support to build upstream package (#2)
What this PR does / why we need it:
Debian upstream discussed to cease support for building google-guest-agent because of it's unmaintained state and as Google Cloud Platform use their own packaging for distribution. Based on apparent removal of packages in Debian testing this PR switches to upstream as well.

Closes #1

Which issue(s) this PR fixes:
Fixes #2621

Signed-off-by: Tobias Wolf <wolf@b1-systems.de>